# feat(rome_js_analyze): faster version of noArguments